[Arc] Add state allocation passes
Add three passes that implement state allocation. The passes take the
abstract state allocation ops, compute a memory layout for the overall
state of the model, and replace the allocation ops with simple pointer
getter ops that access the allocated piece of memory. The passes operate
as follows:

- `LegalizeStateUpdate` detects read-after-write hazards and introduces
  temporary storage locations that allow the read and write ops to
  occur without infringing on each other.
- `AllocateState` computes the overall memory layout and replaces
  allocation ops with simple accessor ops.
- `PrintStateInfo` emits the memory layout as a JSON file. This allows
  other tools to reason about the exact memory layout of the design.

State update legalization is still lacking proper handling of memory
reads and writes, which are significantly more involved than the simple
scalar registers. Follow-up work.

Co-authored-by: Martin Erhart <maerhart@outlook.com>
Co-authored-by: Zachary Yedidia <zyedidia@gmail.com>