[Arc] Add basic LLVM lowering and simple runtime helpers
Add two passes to lower a design to a software model and add utilities
to allow C++ to link against and drive this model.

The `LowerClocksToFuncs` pass outlines all `arc.clock_tree` and
`arc.passthrough` operations into separate MLIR functions. This
conceptually converts clocks from being a signal in the design into a
function that can be called in order to execute the state update
triggered by that clock.

The `LowerToLLVM` pass does exactly as it says: it sets up a dialect
conversion from Arc and the core CIRCT dialects to Func/SCF, and from
there to the LLVM dialect.

Also add an output format option to the arcilator tool that allows for
the direct emission of LLVM IR (as opposed to the MLIR's LLVM dialect).

The `arcilator-header-cpp.py` utility accepts the state JSON file
generated by the `PrintStateInfo` pass and converts it into a C++ header
that makes the lowered LLVM model feel close to Verilator output. The
mapping is not perfect: there is no `eval` function, but dedicated clock
and passthrough functions to be called instead. The commonalities of the
generated C++ header are collected in `arcilator-runtime.h`.

Co-authored-by: Martin Erhart <maerhart@outlook.com>
Co-authored-by: Zachary Yedidia <zyedidia@gmail.com>