[Arc] Add MuxToControlFlow pass
After the muxes that represent enables and resets are factored out by
the `InterStateProperties` pass, there are still a lot of muxes left
that have big fan-ins and can be converted to proper branches for some
performance benefit.

There is still a bit of work to do in the backend, that I haven't
included here, to support lowering of if-statements (or nested regions
in general).

The heuristics for deciding whether to convert to if-statements are not
properly fine-tuned yet because I'm still working on a proper set of
benchmarks to evaluate it and it likely changes anyways until we have a
somewhat stable set of optimizations implemented.